UI: Change workflow "edit all"-button
the button is now relative to the workflowitem list instead of the header
